+++ media.py
@@ -1,4 +1,5 @@
 from baseitem import BaseItem
+import foreign
 
 
 class Media(BaseItem):
@@ -9,4 +10,12 @@ class Media(BaseItem):
         super().__init__(str_id, str_name, str_content, str_link)
         # Here be class vars.
         # str_link hold guid
+        self.excerpt = str_excerpt
+        self.title = str_title
+
+        words = foreign.word_breaker(self.excerpt)
+
+        # Mark foreign words in excerpt
+        self.excerpt_complete_list, self.excerpt_small_list = \
+            foreign.content_lang_marker(words, str_mark_start, str_mark_end, "he")