+++ outputcreator.py
@@ -2,33 +2,35 @@ import post
 import os
 
 
-def text_output(list_of_posts: list, folder_name: str):
+def post_text_output(list_of_items: list, folder_name: str, str_type: str):
     """
     Prints the results to files.
     Small list goes into a single file with ********** separator between posts. File name is 00<folder_name>
     Complete list: Each post gets it's own file
-    :param list_of_posts: All the posts
+    :param str_type: Type of items: Posts and pages or Media
+    :param list_of_items: All the posts
     :param folder_name: Folder to save the files in.
     :return:
     """
     # Check the folder exists
     # Open the small list file for writing and clear it
     folder_maker(folder_name)
-    small_file = open("00"+folder_name+".txt", "w")
+    folder_maker(str_type)
+    small_file = open("00" + folder_name + "-" + str_type + ".txt", "w")
 
     # Iterate on posts, create files for the complete list and append for the small list
-    for single_post in list_of_posts:
+    for single_item in list_of_items:
         # Complete list
-        complete_file = open(single_post.title + ".txt", "w")
+        complete_file = open(single_item.title + ".txt", "w")
         complete_file.write("--------------------------------------------------------------------------------- \n")
-        complete_file.write("Post Title: ... " + single_post.title + "\n")
-        complete_file.write("Post Name: ... " + single_post.name + "\n")
-        complete_file.write("Post Link: ... " + single_post.link + "\n")
+        complete_file.write("Title: ... " + single_item.title + "\n")
+        complete_file.write("Name: ... " + single_item.name + "\n")
+        complete_file.write("Link: ... " + single_item.link + "\n")
         complete_file.write("--------------------------------------------------------------------------------- \n \n")
-        complete_file.write("Post Excerpt: \n")
-        complete_file.write(" ".join(single_post.excerpt_complete_list))
+        complete_file.write("Excerpt: \n")
+        complete_file.write(" ".join(single_item.excerpt_complete_list))
         complete_file.write("--------------------------------------------------------------------------------- \n \n")
-        complete_file.write(" ".join(single_post.content_complete_list))
+        complete_file.write(" ".join(single_item.content_complete_list))
 
         # Close the complete file
         complete_file.close()
@@ -37,17 +39,20 @@ def text_output(list_of_posts: list, folder_name: str):
         small_file.write("************************************************************************************ \n")
         small_file.write("************************************************************************************ \n")
         small_file.write("--------------- \n")
-        small_file.write("Post Title: ... " + single_post.title + "\n")
-        small_file.write("Post Name: ... " + single_post.name + "\n")
-        small_file.write("Post Link: ... " + single_post.link + "\n")
-        small_file.write("Post Excerpt: ... " + " ".join(single_post.excerpt_small_list) + "\n")
+        small_file.write("Title: ... " + single_item.title + "\n")
+        small_file.write("Name: ... " + single_item.name + "\n")
+        small_file.write("Link: ... " + single_item.link + "\n")
+        small_file.write("Excerpt: ... " + " ".join(single_item.excerpt_small_list) + "\n")
         small_file.write("--------------- \n \n")
-        small_file.write(" ".join(single_post.content_small_list) + "\n")
+        small_file.write(" ".join(single_item.content_small_list) + "\n")
         small_file.write("************************************************************************************ \n")
         small_file.write("************************************************************************************ \n")
 
     # Close the small file
     small_file.close()
+
+    # Return to root folder
+    return_to_root()
 
 
 def folder_maker(folder_name: str):
@@ -60,14 +65,21 @@ def folder_maker(folder_name: str):
         os.mkdir(folder_name)
     except FileExistsError:
         pass
-    # else:
-        # print("Fatal error creating the folder")
 
-    # Change path to that dire
+    # Change path to that dir
     os.chdir(folder_name)
 
     # Remove simple list file
     try:
-        os.remove("00"+folder_name)
+        os.remove("00" + folder_name + ".txt")
     except FileNotFoundError:
         pass
+
+
+def return_to_root():
+    """
+    Return pwd to root path
+    :return: none
+    """
+    os.chdir("..")
+    os.chdir("..")