+++ tag.py
@@ -9,7 +9,7 @@ class Tag(BaseItem):
     """
 
     def __init__(self, num_id: str, str_name: str, str_slug: str, str_type: str, str_desc: str, str_mark_start: str,
-                 str_mark_end: str, str_id: str):
+                 str_mark_end: str):
         super().__init__(num_id, str_name, str_desc, str_slug)
         # Here be all class vars
         self.filtered_name = list()  # Name without number, split by -