+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 # v0.2
 1. Less hardcoded settings: string marking, output folder name, 
 2. Excerpt support
-3. Media, tags and h5p support.
+3. Media, Pages, tags and h5p support.
+4. Redesign: Base class for item storage. Tags and Media inherit it. Post and pages in the future.